refactor(adapter): split all report module
Move the all-agent loader, reporting, types, and tests into separate files so the root adapter module stays focused on the public entry points.

Keep the existing row construction and reporting behavior intact, while consolidating the priced summary agent loading helper used by similar adapters to reduce duplicated wrapper code.
